fix: 아지트 멤버 조회, 삭제, 추가를 한 파일로 수정
